moved getContractByType to v1
