+++ docs/_static/auth/v0.yaml
@@ -3,43 +3,7 @@ info:
   title: Nuts Service API Proxy
   version: 0.0.1
 paths:
-  /public/auth/v1/contract/{contractType}:
-    get:
-      operationId: getContractByType
-      summary: Get a contract by type and version
-      tags:
-        - auth
-      parameters:
-        - name: contractType
-          in: path
-          required: true
-          schema:
-            type: string
-        - name: version
-          description: The version of this contract. If omitted, the most recent version will be returned
-          required: false
-          in: query
-          schema:
-            type: string
-        - name: language
-          in: query
-          required: false
-          schema:
-            type: string
-            default: nl
-      responses:
-        '404':
-          description: if no contract exists
-          content:
-            text/plain:
-              schema:
-                type: string
-        '200':
-          description: Returns the contract of this type, version and language
-          content:
-            application/json:
-              schema:
-                $ref: "#/components/schemas/Contract"
+
   /public/auth/v1/accesstoken:
     post:
       operationId: createAccessToken
@@ -144,72 +108,7 @@ paths:
                 $ref: "#/components/schemas/TokenIntrospectionResponse"
 components:
   schemas:
-    ContractType:
-      type: string
-      description: "Type of which contract to sign"
-      example: "BehandelaarLogin"
-    ContractLanguage:
-      type: string
-      description: "Language of the contract in all caps"
-      example: "NL"
-    ContractVersion:
-      type: string
-      description: "Version of the contract"
-      example: "v3"
-    LegalEntity:
-      type: string
-      description: "Identifier of the legalEntity as registered in the Nuts registry"
-      example: "urn:oid:2.16.840.1.113883.2.4.6.1:48000000"
-    Contract:
-      required:
-        - type
-        - version
-        - language
-      properties:
-        type:
-          $ref: "#/components/schemas/ContractType"
-        language:
-          $ref: "#/components/schemas/ContractLanguage"
-        version:
-          $ref: "#/components/schemas/ContractVersion"
-        signer_attributes:
-          example:
-          type: array
-          items:
-            type: string
-        template:
-          type: string
-          example: ik verklaar dat ${acting_party} namens mij request mag maken
-        template_attributes:
-          type: array
-          items:
-            type: string
-          example: [ "irma-demo.MijnOverheid.ageLower.over12",
-                     "irma-demo.MijnOverheid.fullName"
-          ]
-    ContractSigningRequest:
-      required:
-        - type
-        - version
-        - language
-        - legalEntity
-      properties:
-        type:
-          $ref: "#/components/schemas/ContractType"
-        language:
-          $ref: "#/components/schemas/ContractLanguage"
-        version:
-          $ref: "#/components/schemas/ContractVersion"
-        legalEntity:
-          $ref: "#/components/schemas/LegalEntity"
-        valid_from:
-          type: string
-          description: "ValidFrom describes the time from which this contract should be considered valid"
-          example: "2019-06-24T14:32:00+02:00"
-        valid_to:
-          type: string
-          description: "ValidTo describes the time until this contract should be considered valid"
-          example: "2019-12-24T14:32:00+02:00"
+
     CreateAccessTokenRequest:
       description: Request as described in RFC7523 section 2.1
       required:
